+++ cda_api/__init__.py
@@ -1,4 +1,4 @@
-from cda_api.application_utilities import get_logger, get_query_id
+from cda_api.application_functions import get_logger, get_query_id
 from cda_api.classes.exceptions import (
     CDABaseException,
     ColumnNotFound,
