+++ cda_api/db/__init__.py
@@ -1,11 +1,8 @@
 from cda_api import get_logger
-from cda_api.classes.DatabaseMap import DatabaseMap
 from cda_api.classes.DatabaseInfo import DatabaseInfo
-
 from .connection import session
 from .schema import Base
 
-DB_MAP = DatabaseMap(Base)
 DB_INFO = DatabaseInfo(Base)
 log = get_logger("Utility: db/__init__.py")
