--- cda_api/db/query_utilities.py
@@ -1,1026 +1,0 @@
-import itertools
-
-import sqlparse
-from sqlalchemy import CTE, Label, and_, distinct, func, or_, SelectLabelStyle, union_all, union
-from sqlalchemy.exc import CompileError
-from sqlalchemy.dialects import postgresql
-
-from cda_api import RelationshipNotFound, get_logger, TableNotFound
-from cda_api.db import DB_MAP
-
-log = get_logger()
-
-
-# Generates compiled SQL string from query object
-def query_to_string(q, indented=False) -> str:
-    try:
-        sql_string = str(q.statement.compile(compile_kwargs={"literal_binds": True}))
-        if indented:
-            return sqlparse.format(sql_string, reindent=True, keyword_case="upper")
-        else:
-            return sql_string.replace("\n", "")
-    except CompileError as ce:
-        sql_string = str(q.statement.compile())
-        if indented:
-            return sqlparse.format(sql_string, reindent=True, keyword_case="upper")
-        else:
-            return sql_string.replace("\n", "")
-    except Exception as e:
-        raise e
-
-
-# Prints compiled SQL string from query object
-def print_query(q) -> None:
-    print(query_to_string(q, indented=True))
-
-
-# Returns the column object from a cte (Common Table Expression) object
-def get_cte_column(cte, columnname):
-    return getattr(cte.c, columnname)
-
-
-# distinct(count COLUMN)
-def distinct_count(column):
-    return func.count(distinct(column))
-
-def normalize_add_exclude_columns(request_body, endpoint_type, filter_columnnames, log):
-    log.debug("Normalizing [ADD|EXCLUDE]_columns")
-    new_add_columns = filter_columnnames
-    if request_body.ADD_COLUMNS != None:
-        for columnname in request_body.ADD_COLUMNS:
-            if columnname.endswith('.*'):
-                tablename = columnname.replace('.*', '')
-                log.debug(f"{tablename}.* found in ADD_COLUMNS")
-                if endpoint_type == 'data':
-                    log.debug(f"Adding data columns from {tablename} to ADD_COLUMNS")
-                    data_col_infos = DB_MAP.get_table_data_column_infos(tablename) + [col_info for col_info in DB_MAP.get_virtual_table_column_infos(tablename) if col_info.data_returns]
-                    table_columnnames = [col_info.uniquename for col_info in data_col_infos if col_info.uniquename not in new_add_columns]
-                elif endpoint_type == 'summary':
-                    log.debug(f"Adding summary columns from {tablename} to ADD_COLUMNS")
-                    summary_col_infos = DB_MAP.get_table_data_column_infos(tablename) + [col_info for col_info in DB_MAP.get_virtual_table_column_infos(tablename) if col_info.summary_returns]
-                    table_columnnames = [col_info.uniquename for col_info in summary_col_infos if col_info.uniquename not in new_add_columns]
-                else:
-                    raise Exception(f'Unknown endpoint type: "{endpoint_type}". Could not build column list for table: "{tablename}"')
-                new_add_columns = new_add_columns + table_columnnames
-            elif columnname not in new_add_columns:
-                log.debug(f"Adding {columnname} to ADD_COLUMNS")
-                new_add_columns.append(columnname)
-
-    new_exclude_columns = []
-    if request_body.EXCLUDE_COLUMNS != None:
-        for columnname in request_body.EXCLUDE_COLUMNS:
-            if columnname.endswith('.*'):
-                tablename = columnname.replace('.*', '')
-                log.debug(f"{tablename}.* found in EXCLUDE_COLUMNS")
-                if endpoint_type == 'data':
-                    log.debug(f"Adding data columns from {tablename} to EXCLUDE_COLUMNS")
-                    table_columnnames = [col_info.uniquename for col_info in DB_MAP.get_table_data_column_infos(tablename) if col_info.uniquename not in new_exclude_columns]
-                else:
-                    log.debug(f"Adding summary columns from {tablename} to EXCLUDE_COLUMNS")
-                    table_columnnames = [col_info.uniquename for col_info in DB_MAP.get_table_summary_column_infos(tablename) if col_info.uniquename not in new_exclude_columns]
-                new_exclude_columns = new_exclude_columns + table_columnnames
-            elif columnname not in new_exclude_columns:
-                log.debug(f"Adding {columnname} to EXCLUDE_COLUMNS")
-                new_exclude_columns.append(columnname)
-
-    request_body.ADD_COLUMNS = new_add_columns
-    request_body.EXCLUDE_COLUMNS = new_exclude_columns
-
-    return request_body
-
-def apply_match_all_some_filters(query, match_all_filters, match_some_filters, foreign_tablename, virtual_table = None, db=None):
-    if virtual_table == None:
-        log.debug(f"Applying all '{foreign_tablename}' related MATCH_[ALL|SOME] filters to the {foreign_tablename} preselect")
-        # Apply filter conditionals
-        if match_all_filters and match_some_filters:
-            query = query.filter(and_(*match_all_filters)).filter(or_(*match_some_filters))
-        elif match_all_filters:
-            query = query.filter(and_(*match_all_filters))
-        elif match_some_filters:
-            query = query.filter(or_(*match_some_filters))
-        return query
-    else:
-        log.debug(f"Applying all '{virtual_table}' related MATCH_[ALL|SOME] filters to the {foreign_tablename}_[subject|file]_preselect")
-        hanging_table_join = DB_MAP.get_hanging_table_join(
-            hanging_tablename=foreign_tablename, local_tablename=virtual_table
-        )
-        entity_column = list(hanging_table_join["join_table"].foreign_keys)[0].column
-        foreign_column = hanging_table_join["hanging_fk_parent"]
-        # Apply filter conditionals
-        if match_all_filters and match_some_filters:
-            filter_query = db.query(entity_column).filter(and_(*match_all_filters)).filter(or_(*match_some_filters))
-        elif match_all_filters:
-            filter_query = db.query(entity_column).filter(and_(*match_all_filters))
-        elif match_some_filters:
-            filter_query = db.query(entity_column).filter(or_(*match_some_filters))
-        query = query.filter(foreign_column.in_(filter_query))
-        return query
-
-
-# Gets the total distinct counts of a column as a subquery
-def total_column_count_subquery(db, preselect):
-    return db.query(func.count()).select_from(preselect).scalar_subquery()
-    # return db.query(distinct_count(column).label("count_result")).scalar_subquery()
-
-
-# # Gets statistics of a row for numeric columns
-# def numeric_summary(db, column):
-#     column_subquery = db.query(
-#         func.min(column).label("min"),
-#         func.max(column).label("max"),
-#         func.round(func.avg(column)).label("mean"),
-#         func.percentile_disc(0.5).within_group(column).label("median"),
-#         func.percentile_disc(0.25).within_group(column).label("lower_quartile"),
-#         func.percentile_disc(0.75).within_group(column).label("upper_quartile"),
-#     ).subquery("subquery")
-#     # Get the row_to_json of the subquery
-#     column_json = db.query(func.row_to_json(column_subquery.table_valued())
-#                            .label(f"{column.name}_stats")).cte(f"json_{column.name}")
-#     # Apply an array aggregation
-#     numeric_array_agg = db.query(func.array_agg(get_cte_column(column_json, f"{column.name}_stats"))).scalar_subquery()
-#     return numeric_array_agg
-
-
-# Gets the categorical(grouped) json counts of a row
-def categorical_summary(db, column):
-    column_preselect = db.query(column, func.count().label("count_result")).group_by(column).subquery("subquery")
-    # Get the row_to_json of the subquery
-    column_json = db.query(func.row_to_json(column_preselect.table_valued())
-                           .label(f"{column.name}_categories")).cte(f"json_{column.name}")
-    # Apply an array aggregation
-    categorical_array_agg = db.query(func.array_agg(get_cte_column(column_json, f"{column.name}_categories"))).scalar_subquery()
-    return categorical_array_agg
-
-# Gets all combinations of data_source columns
-def get_data_source_combinations(data_source_columnnames):
-    data_source_combinations = {}
-    subsets = itertools.chain(
-        *map(lambda x: itertools.combinations(data_source_columnnames, x), range(0, len(data_source_columnnames) + 1))
-    )
-    for subset in subsets:
-        if len(subset) > 0:
-            combo_boolean = {}
-            name = ""
-            for columnname in data_source_columnnames:
-                boolean = bool(columnname in subset)
-                combo_boolean[columnname] = boolean
-            name = "_".join([name.split("_")[-1] for name, b in combo_boolean.items() if b])
-            if len(subset) < len(data_source_columnnames):
-                name += "_exclusive"
-            data_source_combinations[name] = combo_boolean
-    return data_source_combinations
-
-
-# Combines the counts of data source columns into a single json for use in summary endpoint
-def data_source_counts(db, data_source_columnnames, data_source_columns):
-    data_source_counts = []
-    # Get mapping of all combinations of the data source columns
-    data_source_combinations = get_data_source_combinations(data_source_columnnames)
-    # Get list of queries for each combination of the data_source columns
-    for name, data_source_boolean_map in data_source_combinations.items():
-        filters = [
-            data_source_column == data_source_boolean_map[data_source_column.name]
-            for data_source_column in data_source_columns
-        ]
-        data_source_counts.append(db.query(func.count()).filter(*filters).label(name))
-
-    data_source_preselect = db.query(*data_source_counts).subquery("subquery")
-    # Get the row_to_json of the subquery
-    data_source_json = db.query(func.row_to_json(data_source_preselect.table_valued()))
-    return data_source_json
-
-def get_entity_count_components(db, endpoint_tablename, preselect_query, entity_to_count):
-    entity_relationship = DB_MAP.get_relationship(endpoint_tablename, entity_to_count)
-    entity_local_column = entity_relationship.entity_column
-    entity_local_column_uniquename = DB_MAP.get_column_uniquename(
-        entity_local_column.name, entity_local_column.table.name
-    )
-
-    # Get subquery of the entity_local_column
-    column_uniquename = entity_local_column_uniquename
-    subquery = db.query(
-            get_cte_column(preselect_query, column_uniquename)
-            .label(entity_local_column_uniquename)
-        )
-
-    # If there is a mapping_table (ie. file_describes_subject)
-    if entity_relationship.has_mapping_table:
-        column_to_count = entity_relationship.foreign_mapping_column
-        column_to_filter_on = entity_relationship.entity_mapping_column
-    else:
-        column_to_count = entity_local_column
-        column_to_filter_on = entity_local_column 
-    return subquery, column_to_count, column_to_filter_on
-
-# Gets the total count of an entity's related files and subjects by only counting from the mapping table (ie. observation_of_subject)
-def entity_count(db, endpoint_tablename, preselect_query, entity_to_count, filter_table_map):
-    subquery, column_to_count, column_to_filter_on = get_entity_count_components(db, endpoint_tablename, preselect_query, entity_to_count)
-
-    # Get count subquery
-    entity_count_select = (
-            db.query(func.count(distinct(column_to_count)).label("count_result"))
-            .filter(column_to_filter_on.in_(subquery))
-            .scalar_subquery()
-    )
-    return entity_count_select
-
-# Get array aggregate of unique values in a column not including null
-def unique_column_array_agg(column):
-    return func.array_remove(func.array_agg(distinct(column)), None).label(column.name)
-
-def get_foreign_array_columns_and_join(endpoint_tablename, foreign_tablename):
-    foreign_array_join = None
-    # If there is direct relationship defined
-    if DB_MAP.relationship_exists(endpoint_tablename, foreign_tablename):
-        relation = DB_MAP.get_relationship(endpoint_tablename, foreign_tablename)
-        entity_column = relation.entity_column
-        # If there is a mapping table (ie. file_describes subject)
-        if relation.has_mapping_table:
-            foreign_column = relation.entity_mapping_column
-            foreign_array_join = {
-                'target': relation.mapping_table, 
-                'onclause': relation.foreign_column == relation.foreign_mapping_column
-                }
-
-        # If there is a direct relationship (ie. observation -> subject)
-        else:
-            foreign_column = relation.foreign_column
-
-    # If there is a "hanging table" join defined, (file_tumor_vs_normal is a "hanging table" to file and there is a join defined to subject through file_describes_subject)
-    elif DB_MAP.hanging_table_join_exists(foreign_tablename, endpoint_tablename):
-        hanging_table_join = DB_MAP.get_hanging_table_join(
-            hanging_tablename=foreign_tablename, local_tablename=endpoint_tablename
-        )
-        # If there is a mapping table involved (ie. file_tumor_vs_normal -> "file_describes_subject" -> subject)
-        if "mapping_table_join_clause" in hanging_table_join.keys():
-            entity_column = hanging_table_join["local_column"]
-            foreign_column = hanging_table_join["local_mapping_column"]
-            foreign_array_join = {
-                'target': hanging_table_join["join_table"], 
-                'onclause': hanging_table_join["statement"]
-                }   
-            
-        # If there is a direct connection (ie. file_tumor_vs_normal -> file)
-        else:
-            if foreign_tablename == 'upstream_identifiers':
-                entity_column = DB_MAP.get_meta_column(f'{endpoint_tablename}_id_alias')
-                foreign_column = hanging_table_join["hanging_fk_parent"]
-                # foreign_array_join = {
-                #     'target': hanging_table_join["join_table"], 
-                #     'onclause': entity_column == foreign_column
-                #     }
-            else:
-                entity_column = list(hanging_table_join["join_table"].foreign_keys)[0].column
-                foreign_column = hanging_table_join["hanging_fk_parent"]
-
-    else:
-        error_message = f'Unable to build foreign array preselect between {endpoint_tablename} and {foreign_tablename}'
-        raise RelationshipNotFound(error_message)
-    return entity_column, foreign_column, foreign_array_join
-
-
-# Build the preselects of arrays when adding foreign columns
-def build_foreign_array_preselect(db, endpoint_tablename, foreign_tablename, columns, preselect_query, filter_table_map, log):
-    cte_name = f"{foreign_tablename}_{endpoint_tablename}_columns"
-    log.debug(f"Building {cte_name}")
-    entity_column, foreign_column, foreign_array_join = get_foreign_array_columns_and_join(endpoint_tablename, foreign_tablename)
-    
-    select_cols = [unique_column_array_agg(column).label(column.name) for column in columns] + [foreign_column]
-    foreign_array_preselect = (
-                    db.query(*select_cols)
-                    .filter(foreign_column.in_(preselect_query))
-                    .group_by(foreign_column)
-                )
-    if foreign_tablename in filter_table_map.keys():
-        match_all = filter_table_map[foreign_tablename]['match_all']
-        match_some = filter_table_map[foreign_tablename]['match_some']
-        foreign_array_preselect = apply_match_all_some_filters(foreign_array_preselect, match_all, match_some, foreign_tablename)
-
-    # Check if haning table and apply filter
-    virtual_tables = [DB_MAP.get_column_info(col.name).virtual_table for col in columns]
-    if any(virtual_tables):
-        virtual_table = [table for table in virtual_tables if table != None][0]
-        if virtual_table in filter_table_map.keys():
-            match_all = filter_table_map[virtual_table]['match_all']
-            match_some = filter_table_map[virtual_table]['match_some']
-            foreign_array_preselect = apply_match_all_some_filters(foreign_array_preselect, match_all, match_some, foreign_tablename, virtual_table, db)
-    
-    if foreign_tablename == 'upstream_identifiers':
-        foreign_array_preselect = foreign_array_preselect.filter(DB_MAP.get_meta_column('upstream_identifiers_cda_table') == endpoint_tablename)
-
-    if foreign_array_join:
-        foreign_array_preselect = foreign_array_preselect.join(**foreign_array_join)
-    foreign_array_preselect = foreign_array_preselect.cte(cte_name)
-    
-    onclause = get_cte_column(foreign_array_preselect, foreign_column.name) == entity_column
-    preselect_columns = [col for col in foreign_array_preselect.c if col.name != foreign_column.name]
-    foreign_join = {"target": foreign_array_preselect, "onclause": onclause}
-    return foreign_join, preselect_columns
-
-def build_foreign_json_preselect(db, entity_tablename, foreign_tablename, columns, filter_preselect_query, filter_table_map, log):
-    cte_name = f"{foreign_tablename}_expanded_preselect"
-    log.debug(f"Building {cte_name}")
-    entity_column, entity_mapping_column, foreign_column, foreign_mapping_column, foreign_array_join = get_foreign_array_summary_columns(entity_tablename, foreign_tablename)
-    foreign_column_info = DB_MAP.get_table_column_info(foreign_tablename, foreign_column.name)
-    
-    if (not isinstance(entity_mapping_column, type(None))) and (not isinstance(foreign_mapping_column, type(None))):
-        log.debug(f"Constructing preselect utilizing mapping table")
-        foreign_id_common_name = entity_mapping_column.name
-        foreign_table_filter = foreign_column.in_(db.query(foreign_mapping_column).filter(entity_mapping_column.in_(filter_preselect_query)))
-        foreign_table_columns = [entity_mapping_column.label(entity_mapping_column.name)] + columns
-        foreign_table_subquery = (
-            db.query(*foreign_table_columns)
-            .filter(foreign_table_filter)
-        )
-        if foreign_array_join:
-            foreign_table_subquery = foreign_table_subquery.join(**foreign_array_join)
-    else:
-        log.debug(f"Constructing preselect without a mapping table")
-        foreign_id_common_name = foreign_column_info.columnname
-        foreign_table_columns = [foreign_column.label(foreign_id_common_name)] + columns
-        foreign_table_subquery = (
-            db.query(*foreign_table_columns)
-            .filter(foreign_column.in_(filter_preselect_query))
-        )
-    
-    if foreign_tablename in filter_table_map.keys():
-        log.debug(f"Applying appropriate '{foreign_tablename}' filters to the {foreign_tablename} preselect")
-        match_all = filter_table_map[foreign_tablename]['match_all']
-        match_some = filter_table_map[foreign_tablename]['match_some']
-        foreign_table_subquery = apply_match_all_some_filters(foreign_table_subquery, match_all, match_some, foreign_tablename)
-    
-    # Check if hanging table and apply filter
-    virtual_tables = [DB_MAP.get_column_info(col.name).virtual_table for col in columns]
-    if any(virtual_tables):
-        virtual_table = [table for table in virtual_tables if table != None][0]
-        if virtual_table in filter_table_map.keys():
-            log.debug(f"Applying appropriate '{virtual_table}' filters to the {foreign_tablename} preselect")
-            match_all = filter_table_map[virtual_table]['match_all']
-            match_some = filter_table_map[virtual_table]['match_some']
-            foreign_table_subquery = apply_match_all_some_filters(foreign_table_subquery, match_all, match_some, foreign_tablename, virtual_table, db)
-
-    foreign_table_subquery = foreign_table_subquery.subquery('subquery')
-    
-    if foreign_tablename == 'foreign_json_columns':
-        log.debug(f"Checking if this procs")
-        foreign_table_subquery = foreign_table_subquery.filter(DB_MAP.get_meta_column('upstream_identifiers_cda_table') == entity_tablename)
-
-    log.debug(f"Constructing json_build_object")
-    foreign_json_columns = []
-    subquery_id_column = None
-    for column in foreign_table_subquery.c:
-        if column.name != foreign_id_common_name:
-            foreign_json_columns.append(column.name)
-            foreign_json_columns.append(column)
-        else:
-            subquery_id_column = column
-    
-    json_subquery = db.query(
-                        subquery_id_column.label(foreign_id_common_name),
-                        func.json_build_object(*foreign_json_columns).label('json_results')
-                    ).subquery('json_subquery')
-    
-    json_subquery_id_column = json_subquery.c[foreign_id_common_name]
-
-    log.debug(f"Aggregating rows")
-    foreign_json_preselect = (
-        db.query(
-            json_subquery_id_column.label(foreign_id_common_name),
-            func.array_agg(json_subquery.c['json_results']).label(f'{foreign_tablename}_columns')
-        )
-        .group_by(json_subquery_id_column)
-    )
-    foreign_json_preselect = foreign_json_preselect.cte(cte_name)
-
-    onclause = get_cte_column(foreign_json_preselect, foreign_id_common_name) == entity_column
-    preselect_columns = [foreign_json_preselect.c[f'{foreign_tablename}_columns'].label(f'{foreign_tablename}_columns')]
-    foreign_join = {"target": foreign_json_preselect, "onclause": onclause}
-
-    log.debug(f"Finished building {cte_name}")
-    return foreign_join, preselect_columns
-
-
-def get_foreign_array_summary_subquery(db, entity_tablename, foreign_tablename, select_cols, filter_preselect):
-    entity_column, foreign_column, foreign_array_join = get_foreign_array_columns_and_join(entity_tablename, foreign_tablename)
-    entity_uniquename = DB_MAP.get_column_uniquename(entity_column.name, entity_column.table.name)
-    preselect_column = get_cte_column(filter_preselect, entity_uniquename)
-    filter_array_summary_query = db.query(*select_cols).filter(foreign_column.in_(db.query(preselect_column)))
-
-    if foreign_array_join: 
-        filter_array_summary_query = filter_array_summary_query.join(**foreign_array_join)
-    return filter_array_summary_query.subquery(f"{foreign_tablename}_subquery")
-
-def get_foreign_array_summary_select_columns(db, columns_to_aggregate):
-    select_columns = []
-    for column in columns_to_aggregate:
-        column_count_subquery = (
-            db.query(column, func.count().label("count_result"))
-            .group_by(column)
-            .subquery(f"{column.name}_count_subquery")
-            )
-        column_array = db.query(
-            func.array_agg(
-                func.row_to_json(
-                    column_count_subquery.table_valued()
-                    ).label(f'{column.name}_json')
-                ).label(f'{column.name}_array_agg')
-            ).scalar_subquery()
-        select_columns.append(column_array.label(column.name))
-    return select_columns
-
-
-# def build_foreign_array_summary_preselect(db, entity_tablename, foreign_tablename, columns, preselect_query):
-#     foreign_array_subquery = get_foreign_array_summary_subquery(db, entity_tablename, foreign_tablename, columns, preselect_query)
-
-#     select_columns = get_foreign_array_summary_select_columns(db, foreign_array_subquery.columns)
-
-#     foreign_array_preselect = db.query(*select_columns).cte(f'{foreign_tablename}_columns')
-#     preselect_columns = [col for col in foreign_array_preselect.c]
-    
-#     return foreign_array_preselect, preselect_columns
-def get_foreign_array_summary_columns(entity_tablename, foreign_tablename):
-    foreign_mapping_column = None
-    entity_mapping_column = None
-    foreign_array_join = None
-    # If there is direct relationship defined
-    if DB_MAP.relationship_exists(entity_tablename, foreign_tablename):
-        relation = DB_MAP.get_relationship(entity_tablename, foreign_tablename)
-        entity_column = relation.entity_column
-        entity_mapping_column = relation.entity_mapping_column
-        foreign_column = relation.foreign_column
-        foreign_mapping_column = relation.foreign_mapping_column
-        if relation.has_mapping_table:
-            foreign_array_join = {
-                'target': relation.mapping_table, 
-                'onclause': relation.foreign_column == relation.foreign_mapping_column
-                }
-
-    # If there is a "hanging table" join defined, 
-    # (file_tumor_vs_normal is a "hanging table" to file and there is a join defined to subject through file_describes_subject)
-    elif DB_MAP.hanging_table_join_exists(foreign_tablename, entity_tablename):
-        hanging_table_join = DB_MAP.get_hanging_table_join(
-            hanging_tablename=foreign_tablename, local_tablename=entity_tablename
-        )
-        # If there is a mapping table involved (ie. file_tumor_vs_normal -> "file_describes_subject" -> subject)
-        if "mapping_table_join_clause" in hanging_table_join.keys():
-            entity_column = hanging_table_join["local_column"]
-            entity_mapping_column = hanging_table_join["local_mapping_column"]
-            foreign_column = hanging_table_join["hanging_fk_parent"]
-            foreign_mapping_column = hanging_table_join["foreign_mapping_column"]
-            foreign_array_join = {
-                'target': hanging_table_join["join_table"], 
-                'onclause': hanging_table_join["statement"]
-                }
-        # If there is a direct connection (ie. file_tumor_vs_normal -> file)
-        else:
-            if foreign_tablename == 'upstream_identifiers':
-                entity_column = DB_MAP.get_meta_column(f'{entity_tablename}_id_alias')
-                foreign_column = hanging_table_join["hanging_fk_parent"]
-            else:
-                entity_column = list(hanging_table_join["join_table"].foreign_keys)[0].column
-                foreign_column = hanging_table_join["hanging_fk_parent"]
-
-    else:
-        error_message = f'Unable to build foreign array preselect between {entity_tablename} and {foreign_tablename}'
-        raise RelationshipNotFound(error_message)
-    return entity_column, entity_mapping_column, foreign_column, foreign_mapping_column, foreign_array_join
-    
-def foreign_column_categorical_summary(db, preselect_column, foreign_preselect_id):
-    columnname = preselect_column.name
-    foreign_id_common_name = foreign_preselect_id.name
-    non_null_cte = db.query(preselect_column.label(columnname), foreign_preselect_id.label(foreign_id_common_name)) \
-                    .filter(preselect_column.is_not(None)) \
-                    .group_by(foreign_preselect_id, preselect_column) \
-                    .cte(f'{columnname}_non_nulls')
-    null_cte = db.query(preselect_column.label(columnname), foreign_preselect_id.label(foreign_id_common_name)) \
-            .filter(foreign_preselect_id.not_in(db.query(get_cte_column(non_null_cte, foreign_id_common_name)))) \
-            .group_by(foreign_preselect_id, preselect_column) \
-            .cte(f'{columnname}_nulls')
-    union_subquery = union_all(db.query(non_null_cte.c), db.query(null_cte.c)).set_label_style(SelectLabelStyle.LABEL_STYLE_NONE).subquery(f'{columnname}_union')
-
-
-    union_column = get_cte_column(union_subquery, columnname)
-    count_subquery = db.query(union_column,func.count().label('count_result')) \
-                        .group_by(union_column) \
-                        .subquery(f'{columnname}_count_subquery') 
-
-    column_select = db.query(
-                        func.array_agg(
-                            func.row_to_json(count_subquery.table_valued())
-                        ).label(f'{columnname}_array_agg')).scalar_subquery()
-    return column_select
-
-
-# Gets statistics of a row for numeric columns
-def foreign_column_numeric_summary(db, preselect_column):
-    column_subquery = db.query(
-        func.min(preselect_column).label("min"),
-        func.max(preselect_column).label("max"),
-        func.round(func.avg(preselect_column)).label("mean"),
-        func.percentile_disc(0.5).within_group(preselect_column).label("median"),
-        func.percentile_disc(0.25).within_group(preselect_column).label("lower_quartile"),
-        func.percentile_disc(0.75).within_group(preselect_column).label("upper_quartile"),
-    ).subquery("subquery")
-    # Get the row_to_json of the subquery
-    column_json = db.query(func.row_to_json(column_subquery.table_valued())
-                           .label(f"{preselect_column.name}_stats")).cte(f"json_{preselect_column.name}")
-    # Apply an array aggregation
-    numeric_array_agg = db.query(func.array_agg(get_cte_column(column_json, f"{preselect_column.name}_stats"))).scalar_subquery()
-    return numeric_array_agg
-
-
-def build_foreign_array_summary_preselect(db, endpoint_tablename, foreign_tablename, columns, preselect_query, filter_table_map):
-    entity_column, entity_mapping_column, foreign_column, foreign_mapping_column, _= get_foreign_array_summary_columns(endpoint_tablename, foreign_tablename)
-    entity_column_info = DB_MAP.get_table_column_info(endpoint_tablename, entity_column.name)
-    foreign_column_info = DB_MAP.get_table_column_info(foreign_tablename, foreign_column.name)
-    foreign_id_common_name = foreign_column_info.columnname
-    if foreign_id_common_name == 'id_alias':
-        foreign_id_common_name = foreign_mapping_column.name
-    preselect_entity_column = db.query(get_cte_column(preselect_query, entity_column_info.uniquename).label(foreign_id_common_name))
-    if (not isinstance(entity_mapping_column, type(None))) and (not isinstance(foreign_mapping_column, type(None))):
-        foreign_table_id_cte = db.query(foreign_mapping_column) \
-                                .filter(entity_mapping_column.in_(preselect_entity_column))
-    else:
-        foreign_table_id_cte = preselect_entity_column
-    foreign_table_id_cte = foreign_table_id_cte.cte(f'{foreign_tablename}_mapping_ids')
-    foreign_table_id_cte_column = db.query(get_cte_column(foreign_table_id_cte, foreign_id_common_name))
-    
-    foreign_columns = [foreign_column.label(foreign_id_common_name)] + [DB_MAP.get_column_info(column.name).labeled_column for column in columns]
-    foreign_table_preselect = db.query(*foreign_columns).filter(foreign_column.in_(foreign_table_id_cte_column))
-
-    if foreign_tablename in filter_table_map.keys():
-        match_all = filter_table_map[foreign_tablename]['match_all']
-        match_some = filter_table_map[foreign_tablename]['match_some']
-        foreign_table_preselect = apply_match_all_some_filters(foreign_table_preselect, match_all, match_some, foreign_tablename)
-    
-    foreign_table_preselect = foreign_table_preselect.cte(f'{foreign_tablename}_preselect')
-    foreign_preselect_id = get_cte_column(foreign_table_preselect, foreign_id_common_name)
-
-    # Get file or subject count
-    if endpoint_tablename != 'subject':
-        entity_to_count = 'subject'
-    else:
-        entity_to_count = 'file'
-    
-    entity_total_count_select = None
-    if foreign_tablename == entity_to_count:
-        entity_total_count_select = db.query(func.count(distinct(foreign_preselect_id))).scalar_subquery().label(f'{entity_to_count}_count')
-    
-    foreign_selects = []
-    for column in columns:
-        column_info = DB_MAP.get_column_info(column.name)
-        preselect_column = get_cte_column(foreign_table_preselect, column.name)
-        if not column_info.summary_returns:
-            log.debug(f'Skipping column: {column_info.uniquename} because it is not supposed to be displayed')
-            continue
-        if column_info.process_before_display != 'data_source':  
-            match column_info.column_type:
-                case 'numeric':
-                    column_summary = foreign_column_numeric_summary(db, preselect_column)
-                    foreign_selects.append(column_summary.label(column_info.uniquename))
-                case 'categorical':
-                    column_summary = foreign_column_categorical_summary(db, preselect_column, foreign_preselect_id)
-                    foreign_selects.append(column_summary.label(column_info.uniquename))
-                case _:
-                    log.warning(f'Unexpectedly skipping {column_info.uniquename} for summary - column_type: {column_info.column_type}')
-                    pass
-
-    foreign_array_preselect = db.query(*foreign_selects).cte(f'{foreign_tablename}_columns')
-    return foreign_array_preselect.c, entity_total_count_select
-    
-def get_foreign_array_summary_selects(db, endpoint_tablename, add_columns, preselect_query, filter_table_map, log):
-    summary_selects = []
-    foreign_array_map = {}
-    for columnname in add_columns:
-        column_info = DB_MAP.get_column_info(columnname)
-        if column_info.tablename != endpoint_tablename:
-            if column_info.tablename not in foreign_array_map.keys():
-                foreign_array_map[column_info.tablename] = [column_info.metadata_column.label(column_info.uniquename)]
-            else:
-                foreign_array_map[column_info.tablename].append(column_info.metadata_column.label(column_info.uniquename))
-    entity_total_count_select = None
-    for foreign_tablename, columns in foreign_array_map.items():
-        # build_foreign_array_summary_preselect(db, endpoint_tablename, foreign_tablename, columns, preselect_query, filter_table_map)
-        preselect_columns, possible_entity_total_count_select = build_foreign_array_summary_preselect(db, endpoint_tablename, foreign_tablename, columns, preselect_query, filter_table_map)
-        if not isinstance(possible_entity_total_count_select, type(None)):
-            entity_total_count_select = possible_entity_total_count_select
-        for column in preselect_columns:
-            summary_selects.append(db.query(column).label(f'{column.name}_summary'))
-    return summary_selects, entity_total_count_select
-
-
-
-def build_filter_preselect(db, endpoint_tablename, match_all_conditions, match_some_conditions, log):
-    log.debug(f"Building filter preselect")
-    # Get the id_alias column
-    endpoint_id_alias = DB_MAP.get_meta_column(f"{endpoint_tablename}_id_alias")
-
-    # Set up the CTE preselect by selecting the id_alias column from it
-    preselect_cte = db.query(endpoint_id_alias.label("id_alias"))
-
-    preselect_cte = apply_match_all_some_filters(preselect_cte, match_all_conditions, match_some_conditions, endpoint_tablename)
-
-    preselect_cte = preselect_cte.cte(f"{endpoint_tablename}_preselect")
-    preselect_query = db.query(preselect_cte.c.id_alias)
-    return preselect_query, endpoint_id_alias
-
-
-# Combines select columns, match conditions, and mapping columns into cohesive query
-def build_match_query(db, select_columns, match_all_conditions=None, match_some_conditions=None, mapping_joins=None):
-    # Add select columns
-    query = db.query(*select_columns)
-
-    # Add filters
-    if match_all_conditions and match_some_conditions:
-        query = query.filter(and_(*match_all_conditions)).filter(or_(*match_some_conditions))
-    elif match_all_conditions:
-        query = query.filter(and_(*match_all_conditions))
-    elif match_some_conditions:
-        query = query.filter(or_(*match_some_conditions))
-
-    # Add joins
-    if mapping_joins:
-        # Add joins individually since .join(*mapping_columns) doesn't work for some reason
-        for mapping_join in mapping_joins:
-            query = query.join(**mapping_join)
-
-    return query
-
-def get_hanging_table_join(endpoint_tablename, select_column):
-    if isinstance(select_column, Label):
-        column_table = select_column.element.table
-        if not isinstance(column_table, CTE):
-            column_tablename = column_table.name
-            if column_tablename in DB_MAP.hanging_table_relationship_map.keys():
-                if endpoint_tablename in DB_MAP.hanging_table_relationship_map[column_tablename].keys():
-                    hanging_table_join = DB_MAP.get_hanging_table_join(
-                        hanging_tablename=column_tablename, local_tablename=endpoint_tablename
-                    )
-                    return {'target': hanging_table_join["join_table"], 'onclause': hanging_table_join["statement"]}
-    return None
-    
-
-def add_hanging_table_joins(endpoint_tablename, select_columns, query):
-    hanging_tablenames = []
-    for column in select_columns:
-        if isinstance(column, Label):
-            column_table = column.element.table
-            if not isinstance(column_table, CTE):
-                column_tablename = column_table.name
-                if (
-                    column_tablename in DB_MAP.hanging_table_relationship_map.keys()
-                    and column_tablename not in hanging_tablenames
-                ):
-                    hanging_tablenames.append(column_table.name)
-
-    for hanging_tablename in hanging_tablenames:
-        if endpoint_tablename in DB_MAP.hanging_table_relationship_map[hanging_tablename].keys():
-            hanging_table_join = DB_MAP.get_hanging_table_join(
-                hanging_tablename=hanging_tablename, local_tablename=endpoint_tablename
-            )
-            query = query.join(hanging_table_join["join_table"], hanging_table_join["statement"])
-        else:
-            log.warning(f"Unable to map {hanging_tablename} and {endpoint_tablename}")
-    return query
-
-
-def get_identifiers_preselect_columns(db, entity_tablename, preselect_query, log):
-    log.debug(f"Building {entity_tablename}_identifiers select")
-    ui_cda_table_info = DB_MAP.get_column_info('upstream_identifiers_cda_table')
-    ui_id_alias_info = DB_MAP.get_column_info('upstream_identifiers_id_alias')
-    ui_data_source_info = DB_MAP.get_column_info('upstream_identifiers_data_source')
-    ui_data_source_id_field_name_info = DB_MAP.get_column_info('data_source_id_field_name')
-    ui_data_source_id_value_info = DB_MAP.get_column_info('data_source_id_value')
-    ui_subquery = (
-        db.query(
-            ui_id_alias_info.metadata_column.label(ui_id_alias_info.uniquename),
-            ui_data_source_info.metadata_column.label('data_source'),
-            ui_data_source_id_field_name_info.metadata_column.label(ui_data_source_id_field_name_info.uniquename),
-            ui_data_source_id_value_info.metadata_column.label(ui_data_source_id_value_info.uniquename)
-        )
-        .filter(ui_cda_table_info.metadata_column == entity_tablename)
-        .filter(ui_id_alias_info.metadata_column.in_(preselect_query))
-        .subquery('subquery')
-    )
-    ui_json_subquery = (
-        db.query(
-            ui_subquery.c[ui_id_alias_info.uniquename].label('id_alias'),
-            func.json_build_object(
-                'data_source', ui_subquery.c['data_source'],
-                ui_subquery.c[ui_data_source_id_field_name_info.uniquename], ui_subquery.c[ui_data_source_id_value_info.uniquename]
-            ).label('json_results')
-        )
-        .subquery('json_subquery')
-    )
-    ui_preselect = (
-        db.query(
-            ui_json_subquery.c['id_alias'].label('id_alias'),
-            func.array_agg(ui_json_subquery.c['json_results']).label(f'{entity_tablename}_identifiers')
-        )
-        .group_by(ui_json_subquery.c['id_alias'])
-    )
-    ui_preselect = ui_preselect.cte(f'{entity_tablename}_identifiers_preselect')
-    preselect_columns = [get_cte_column(ui_preselect, f'{entity_tablename}_identifiers')]
-    onclause = get_cte_column(ui_preselect, f'id_alias') == DB_MAP.get_meta_column(f'{entity_tablename}_id_alias')
-    foreign_join = {"target": ui_preselect, "onclause": onclause}
-
-    return foreign_join, preselect_columns
-    
-    
-def apply_match_all_and_some_filters(query, match_all_db_filters, match_some_db_filters):
-    if match_all_db_filters and match_some_db_filters:
-        query = query.filter(and_(*match_all_db_filters)).filter(or_(*match_some_db_filters))
-    elif match_all_db_filters:
-        query = query.filter(and_(*match_all_db_filters))
-    elif match_some_db_filters:
-        query = query.filter(or_(*match_some_db_filters))
-    return query
-
-
-def build_virtual_foreign_arrays(db,foreign_table_info, virtual_table_info, virtual_column_infos, filtered_preselect, log):
-    virtual_table_relationship = foreign_table_info.get_table_relationship(virtual_table_info)
-    if virtual_table_relationship.requires_mapping_table:
-        relating_column = virtual_table_relationship.local_mapping_column_info.db_column
-    else:
-        relating_column = virtual_table_relationship.foreign_column_info.db_column
-    cte_name = f'{foreign_table_info.name}_{virtual_table_info.name}_columns'
-    select_columns = [relating_column] + [unique_column_array_agg(column_info.db_column).label(column_info.name) for column_info in virtual_column_infos]
-    virtual_preselect = (
-        db.query(*select_columns)
-        .filter(relating_column.in_(filtered_preselect))
-    )
-    virtual_preselect = virtual_preselect.group_by(relating_column)
-    virtual_preselect_cte = virtual_preselect.cte(cte_name)
-    preselect_onclause = get_cte_column(virtual_preselect_cte, relating_column.name) == virtual_table_relationship.local_column_info.db_column
-    preselect_join = {'target': virtual_preselect_cte, 'onclause': preselect_onclause}
-    preselect_columns = [db_column for db_column in virtual_preselect_cte.c if db_column.name != relating_column.name]
-    return preselect_columns, [preselect_join]
-
-
-
-
-def build_foreign_preselect(construct_type, db, endpoint_table_info, relating_table_info, filtered_preselect, foreign_table_info, column_infos, foreign_filter_infos, log):
-    log.debug(f'Building foreign array or {foreign_table_info} connecting to {endpoint_table_info} through {relating_table_info}')
-    # Get relationship info
-    if relating_table_info != foreign_table_info:
-        filter_relationship = relating_table_info.get_table_relationship(foreign_table_info)
-        if filter_relationship.requires_mapping_table:
-            filtered_preselect_relating_column = filter_relationship.local_mapping_column_info.db_column
-        else:
-            filtered_preselect_relating_column = filter_relationship.foreign_column_info.db_column
-    else:
-        filtered_preselect_relating_column = relating_table_info.primary_key_column_info.db_column
-
-    endpoint_relationship = endpoint_table_info.get_table_relationship(foreign_table_info)
-    if endpoint_relationship.requires_mapping_table:
-        endpoint_relating_column = endpoint_relationship.local_mapping_column_info.db_column
-    else:
-        endpoint_relating_column = endpoint_relationship.foreign_column_info.db_column
-
-    virtual_column_info_map = {}
-    foreign_column_infos = []
-    for column_info in column_infos:
-        if column_info.parent_table_info != foreign_table_info:
-            virtual_table_info = column_info.parent_table_info
-            if virtual_table_info not in virtual_column_info_map.keys():
-                virtual_column_info_map[virtual_table_info] = []
-            virtual_column_info_map[virtual_table_info].append(column_info)
-        else:
-            foreign_column_infos.append(column_info)
-
-    # Name and select columns slightly differ between json & array
-    if construct_type == 'json':
-        cte_name = f"{foreign_table_info.name}_expanded_preselect"
-        select_columns = [endpoint_relating_column] + [column_info.labeled_db_column for column_info in foreign_column_infos]
-    elif construct_type == 'array':
-        cte_name = f'{foreign_table_info.name}_{endpoint_table_info.name}_columns'
-        select_columns = [endpoint_relating_column] + [unique_column_array_agg(column_info.db_column).label(column_info.name) for column_info in foreign_column_infos]
-    elif construct_type == 'provenance':
-        cte_name = f'{relating_table_info.name}_identifiers_preselect'
-        select_columns = [endpoint_relating_column] + [column_info.original_labeled_column for column_info in foreign_column_infos]
-    else:
-        raise Exception(f'Unexpected foreign preselect contruct_type {construct_type}. Please use "json", or "array"')
-    
-    virtual_table_joins = []
-    if construct_type == 'array':
-        for virtual_table_info, virtual_column_infos in virtual_column_info_map.items():
-            select_columns += [unique_column_array_agg(column_info.db_column).label(column_info.name) for column_info in virtual_column_infos]
-            virtual_table_relationship = foreign_table_info.get_table_relationship(virtual_table_info)
-            virtual_table_joins.append(virtual_table_relationship.get_foreign_table_join_clause())
-    else:
-        for virtual_table_info, virtual_column_infos in virtual_column_info_map.items():
-            columns, joins = build_virtual_foreign_arrays(db, foreign_table_info, virtual_table_info, virtual_column_infos, filtered_preselect, log)
-            if construct_type == 'json':
-                for col in columns:
-                    select_columns.append(func.coalesce(col, []).label(col.name))
-            else:
-                select_columns.extend(columns)
-            virtual_table_joins.extend(joins)
-
-    log.debug(f"Building {cte_name}")
-
-
-    # Set up base preselect
-    foreign_preselect = (
-        db.query(*select_columns)
-        .filter(filtered_preselect_relating_column.in_(filtered_preselect))
-    )
-
-    for virtual_table_join in virtual_table_joins:
-        foreign_preselect = foreign_preselect.join(**virtual_table_join, isouter=True)
-
-    # Join on the mapping table if required
-    if endpoint_relationship.requires_mapping_table:
-        foreign_preselect = foreign_preselect.join(**endpoint_relationship.get_foreign_table_join_clause())
-    
-    # Apply additional filters if required
-    for additional_filter in endpoint_relationship.additional_filters:
-        foreign_preselect = foreign_preselect.filter(additional_filter)
-
-    # Add filters if the current foreign table is file or subject
-    for filter_info in foreign_filter_infos:
-        if filter_info.filter_column_info.parent_table_info in relating_table_info.database_info.local_table_infos:
-            foreign_preselect = foreign_preselect.filter(filter_info.get_db_filter_for_table(foreign_table_info))
-
-
-    # Build individual column arrays
-    if construct_type == 'array': 
-        foreign_preselect = foreign_preselect.group_by(endpoint_relating_column)
-        foreign_array_preselect_cte = foreign_preselect.cte(cte_name)
-        preselect_onclause = get_cte_column(foreign_array_preselect_cte, endpoint_relating_column.name) == endpoint_relationship.local_column_info.db_column
-        preselect_join = {'target': foreign_array_preselect_cte, 'onclause': preselect_onclause}
-        preselect_columns = [db_column for db_column in foreign_array_preselect_cte.c if db_column.name != endpoint_relating_column.name]
-        
-    # Build json for all columns
-    elif construct_type == 'json': 
-        # Need to auto add data_at and data_source columns
-        foreign_table_subquery = foreign_preselect.subquery('subquery')
-        foreign_json_columns = []
-        subquery_id_column = None
-        for column in foreign_table_subquery.c:
-            if column.name != endpoint_relating_column.name:
-                foreign_json_columns.append(column.name)
-                foreign_json_columns.append(column)
-            else:
-                subquery_id_column = column
-        
-        json_subquery = db.query(
-                            subquery_id_column.label(endpoint_relating_column.name),
-                            func.json_build_object(*foreign_json_columns).label('json_results')
-                        ).subquery('json_subquery')
-        
-        json_subquery_id_column = json_subquery.c[endpoint_relating_column.name]
-
-        log.debug(f"Aggregating rows")
-        foreign_json_preselect = (
-            db.query(
-                json_subquery_id_column.label(endpoint_relating_column.name),
-                func.array_agg(json_subquery.c['json_results']).label(f'{foreign_table_info.name}_columns')
-            )
-            .group_by(json_subquery_id_column)
-        )
-        foreign_json_preselect = foreign_json_preselect.cte(cte_name)
-
-        onclause = get_cte_column(foreign_json_preselect, endpoint_relating_column.name) == endpoint_relationship.local_column_info.db_column
-        preselect_columns = [foreign_json_preselect.c[f'{foreign_table_info.name}_columns'].label(f'{foreign_table_info.name}_columns')]
-        preselect_join = {"target": foreign_json_preselect, "onclause": onclause}
-
-    else:
-        ui_subquery = foreign_preselect.subquery('subquery')
-        ui_json_subquery = (
-            db.query(
-                ui_subquery.c[endpoint_relating_column.name].label('id_alias'),
-                func.json_build_object(
-                    'data_source', ui_subquery.c['data_source'],
-                    ui_subquery.c['data_source_id_field_name'], ui_subquery.c['data_source_id_value']
-                ).label('json_results')
-            )
-            .subquery('json_subquery')
-        )
-        ui_preselect = (
-            db.query(
-                ui_json_subquery.c['id_alias'].label('id_alias'),
-                func.array_agg(ui_json_subquery.c['json_results']).label(f'{relating_table_info.name}_identifiers')
-            )
-            .group_by(ui_json_subquery.c['id_alias'])
-        )
-        ui_preselect = ui_preselect.cte(cte_name)
-        preselect_columns = [get_cte_column(ui_preselect, f'{relating_table_info.name}_identifiers')]
-        onclause = get_cte_column(ui_preselect, f'id_alias') == relating_table_info.primary_key_column_info.db_column
-        preselect_join = {"target": ui_preselect, "onclause": onclause}
-
-    log.debug(f"Finished building {cte_name}")
-    return preselect_columns, [preselect_join]
-
-
-
-# Gets the total distinct counts of a column as a subquery
-def column_distinct_count_subquery(db, column):
-    return db.query(func.count(distinct(column))).scalar_subquery()
-
-def foreign_table_distinct_count(db, endpoint_preselect, endpoint_table_info, foreign_table_info):
-    table_relationship = endpoint_table_info.get_table_relationship(foreign_table_info)
-    if table_relationship.requires_mapping_table:
-        column_to_count = table_relationship.foreign_mapping_column_info.db_column
-        column_to_filter = table_relationship.local_mapping_column_info.db_column
-    else:
-        column_to_count = table_relationship.foreign_column_info.db_column
-        column_to_filter = table_relationship.local_column_info
-
-
-    # Get count subquery
-    entity_count_select = (
-            db.query(func.count(distinct(column_to_count)).label("count_result"))
-            .filter(column_to_filter.in_(endpoint_preselect))
-            .scalar_subquery()
-    )
-    return entity_count_select
-
-
-# Gets statistics of a row for numeric columns
-def numeric_summary(db, column):
-    column_subquery = db.query(
-        func.min(column).label("min"),
-        func.max(column).label("max"),
-        func.round(func.avg(column)).label("mean"),
-        func.percentile_disc(0.5).within_group(column).label("median"),
-        func.percentile_disc(0.25).within_group(column).label("lower_quartile"),
-        func.percentile_disc(0.75).within_group(column).label("upper_quartile"),
-    ).subquery("subquery")
-    # Get the row_to_json of the subquery
-    column_json = db.query(func.row_to_json(column_subquery.table_valued())
-                           .label(f"{column.name}_stats")).cte(f"json_{column.name}")
-    # Apply an array aggregation
-    numeric_array_agg = db.query(func.array_agg(get_cte_column(column_json, f"{column.name}_stats"))).scalar_subquery()
-    return numeric_array_agg
-
-
-# Gets the categorical(grouped) json counts of a row
-def basic_categorical_summary(db, column):
-    column_preselect = db.query(column, func.count().label("count_result")).group_by(column).subquery("subquery")
-    # Get the row_to_json of the subquery
-    column_json = db.query(func.row_to_json(column_preselect.table_valued())
-                           .label(f"{column.name}_categories")).cte(f"json_{column.name}")
-    # Apply an array aggregation
-    categorical_array_agg = db.query(func.array_agg(get_cte_column(column_json, f"{column.name}_categories"))).scalar_subquery()
-    return categorical_array_agg
-
-def null_aware_categorical_summary(db, db_column, connecting_column):
-    non_null_cte = db.query(db_column, connecting_column) \
-                    .filter(db_column.is_not(None)) \
-                    .group_by(connecting_column, db_column) \
-                    .cte(f'{db_column.name}_non_nulls')
-    null_cte = db.query(db_column, connecting_column) \
-                    .filter(connecting_column.not_in(db.query(get_cte_column(non_null_cte, connecting_column.name)))) \
-                    .group_by(connecting_column, db_column) \
-                    .cte(f'{db_column.name}_nulls')
-    union_subquery = union_all(db.query(non_null_cte.c), db.query(null_cte.c)).set_label_style(SelectLabelStyle.LABEL_STYLE_NONE).subquery(f'{db_column.name}_union')
-    union_column = get_cte_column(union_subquery, db_column.name)
-    count_subquery = db.query(union_column,func.count().label('count_result')) \
-                        .group_by(union_column) \
-                        .subquery(f'{db_column.name}_count_subquery') 
-
-    categorical_array_agg = db.query(
-                        func.array_agg(
-                            func.row_to_json(count_subquery.table_valued().label(f'{db_column.name}_categories'))
-                        ).label(f'{db_column.name}_array_agg')).scalar_subquery()
-    return categorical_array_agg
-
-# # Gets all combinations of data_source columns
-# def get_data_source_combinations(data_source_columnnames):
-#     data_source_combinations = {}
-#     subsets = itertools.chain(
-#         *map(lambda x: itertools.combinations(data_source_columnnames, x), range(0, len(data_source_columnnames) + 1))
-#     )
-#     for subset in subsets:
-#         if len(subset) > 0:
-#             combo_boolean = {}
-#             name = ""
-#             for columnname in data_source_columnnames:
-#                 boolean = bool(columnname in subset)
-#                 combo_boolean[columnname] = boolean
-#             name = "_".join([name.split("_")[-1] for name, b in combo_boolean.items() if b])
-#             if len(subset) < len(data_source_columnnames):
-#                 name += "_exclusive"
-#             data_source_combinations[name] = combo_boolean
-#     return data_source_combinations
-
-
-# # Combines the counts of data source columns into a single json for use in summary endpoint
-# def data_source_counts(db, data_source_columns):
-#     data_source_counts = []
-#     data_source_columnnames = [column.name for column in data_source_columns]
-#     # Get mapping of all combinations of the data source columns
-#     data_source_combinations = get_data_source_combinations(data_source_columnnames)
-#     # Get list of queries for each combination of the data_source columns
-#     for name, data_source_boolean_map in data_source_combinations.items():
-#         filters = [
-#             data_source_column == data_source_boolean_map[data_source_column.name]
-#             for data_source_column in data_source_columns
-#         ]
-#         data_source_counts.append(db.query(func.count()).filter(*filters).label(name))
-
-#     data_source_preselect = db.query(*data_source_counts).subquery("subquery")
-#     # Get the row_to_json of the subquery
-#     data_source_json = db.query(func.row_to_json(data_source_preselect.table_valued()))
-#     return data_source_json
