+++ cda_api/routers/data.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException, Request
 from sqlalchemy.orm import Session
 
 from cda_api import EmptyQueryError, get_logger, get_query_id
-from cda_api.application_utilities import handle_router_errors
+from cda_api.application_functions import handle_router_errors
 from cda_api.db import get_db
 from cda_api.db.query_builders import data_query
 from cda_api.models import PagedResponseObj, DataRequestBody
