+++ cda_api/routers/summary.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException, Request
 from sqlalchemy.orm import Session
 
 from cda_api import EmptyQueryError, get_logger, get_query_id
-from cda_api.application_utilities import handle_router_errors
+from cda_api.application_functions import handle_router_errors
 from cda_api.db import get_db
 from cda_api.db.query_builders import summary_query
 from cda_api.models import SummaryResponseObj, SummaryRequestBody
@@ -34,7 +34,7 @@ def file_summary_endpoint(request: Request, request_body: SummaryRequestBody, db
         raise HTTPException(status_code=404, detail=str(e))
 
     try:
-        result = summary_query(db, endpoint_tablename="file", request_body=request_body, log=log)
+        result = summary_query(db, endpoint_table_name="file", request_body=request_body, log=log)
         log.info("Success")
     except Exception as e:
         handle_router_errors(e, log)
@@ -65,7 +65,7 @@ def subject_summary_endpoint(request: Request, request_body: SummaryRequestBody,
         raise HTTPException(status_code=404, detail=str(e))
 
     try:
-        result = summary_query(db, endpoint_tablename="subject", request_body=request_body, log=log)
+        result = summary_query(db, endpoint_table_name="subject", request_body=request_body, log=log)
         log.info("Success")
     except Exception as e:
         handle_router_errors(e, log)
